reorganize code with review feedback
